+++ src/ts/VAST/Parsers/ProgrammaticVastParser.ts
@@ -89,7 +89,6 @@ export class ProgrammaticVastParser extends CampaignParser {
             session: session,
             mediaId: response.getMediaId(),
             trackingUrls: response.getTrackingUrls() || {},
-            backupCampaign: false,
             isLoadEnabled: false
         };